test: share CLIServe cache clock setup
